refactor: user model
