+++ frontend/lib/features/home/repository/profile_repository.dart
@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 
 import 'package:fundflow/app.dart';
 
-import 'package:fundflow/features/setting/models/user_profile.dart';
+import 'package:fundflow/models/user_model.dart';
 import 'package:fundflow/utils/api_helper.dart';
 
 class ProfileRepository {
@@ -10,11 +10,11 @@ class ProfileRepository {
 
   ProfileRepository({required ApiHelper apiHelper}) : dio = apiHelper.dio;
 
-  Future<UserProfile> getUserProfile() async {
+  Future<User> getUserProfile() async {
     try {
       final response = await dio.get('/profile');
       logger.d('Fetched user profile: ${response.data}');
-      return UserProfile.fromJson(response.data);
+      return User.fromJson(response.data);
     } on DioException catch (e) {
       logger.e('Failed to fetch user profile: ${e.response?.data}');
       throw Exception('Failed to fetch user profile');
