+++ frontend/lib/models/user_model.dart
